+++ CHANGES.md
@@ -3,6 +3,7 @@
 - Don't anchor the cursor to the character in combat
 - Guard a bunch of our cursor related keys to make it clear that they are unavailable if combat mode is on
 - Fix: don't cursor skip in combat mode if the player's hand does not contain a capsule. Instead, error.
+- Fix: shift+r now rotates outside combat mode.
 
 # 0.16.38 (2025-12-16)